Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/recorder/keyboard.c
@@ -255,7 +255,7 @@ static void kbd_inschar(struct edit_state *state, unsigned short ch)
 
     len = strlen(state->text);
     utf8 = utf8encode(ch, tmp);
-    j = (long)utf8 - (long)tmp;
+    j = (intptr_t)utf8 - (intptr_t)tmp;
 
     if (len + j < state->buflen)
     {
